Avoid exporting things from generator code.
